reorganize e2e code helpers
